Fixed radius for dungeon template imports

+++ src/eve-common/EVE_Dungeon.h
@@ -188,7 +188,7 @@ namespace Dungeon {
         double yaw;
         double pitch;
         double roll;
-        uint16 radius;
+        double radius;
     };
 
     struct RoomData {
